Refactored extraction into ReleaseFetcher
