+++ src/Factoriod.Fetcher/ReleaseFetcher.cs
@@ -3,6 +3,8 @@ using System.IO;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using SharpCompress.Common;
+using SharpCompress.Readers;
 
 namespace Factoriod.Fetcher
 {
@@ -15,32 +17,56 @@ namespace Factoriod.Fetcher
             this.client = client;
         }
 
-        public async Task<FileInfo> DownloadToAsync(FactorioVersion version, Distro distro, DirectoryInfo outputDirectory, CancellationToken cancellationToken = default)
+        public async Task<DirectoryInfo?> DownloadToAsync(FactorioVersion version, Distro distro, DirectoryInfo outputDirectory, CancellationToken cancellationToken = default)
         {
             var versionedOutputDirectory = Path.Join(outputDirectory.FullName, version.Version.ToString(), version.Build.ToString(), Distro.Linux64.ToString());
             Directory.CreateDirectory(versionedOutputDirectory);
             var outputFile = new FileInfo(Path.Join(versionedOutputDirectory, "factorio.tar.xz"));
+
             Console.WriteLine($"Downloading to {outputDirectory}");
 
-            if (outputFile.Exists)
+            if (!outputFile.Exists)
             {
-                Console.WriteLine($"File already exists: {outputFile}");
-                return outputFile;
+                Console.WriteLine($"File does not exists, downloading");
+                var downloadUrl = GetDownloadUrl(version, distro);
+                using var response = await client.GetAsync(downloadUrl, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
+                response.EnsureSuccessStatusCode();
+
+                using var requestStream = await response.Content.ReadAsStreamAsync(cancellationToken);
+                using var outputFileStream = File.Open(outputFile.FullName, FileMode.Create);
+                await requestStream.CopyToAsync(outputFileStream, cancellationToken);
+                response.Content = null;
             }
 
-            var downloadUrl = GetDownloadUrl(version, distro);
-            using var response = await client.GetAsync(downloadUrl, HttpCompletionOption.ResponseHeadersRead, cancellationToken);
-            response.EnsureSuccessStatusCode();
+            if (outputFile.Directory == null)
+            {
+                Console.WriteLine($"Could not find directory of {outputFile}");
+                return null;
+            }
 
-            using var requestStream = await response.Content.ReadAsStreamAsync(cancellationToken);
-            using var outputFileStream = File.Open(outputFile.FullName, FileMode.Create);
-            await requestStream.CopyToAsync(outputFileStream, cancellationToken);
-            response.Content = null;
-
-            return outputFile;
+            return Extract(outputFile, outputFile.Directory);
         }
 
         private static string GetDownloadUrl(FactorioVersion version, Distro distro)
             => $"https://factorio.com/get-download/{version.Version}/{version.Build}/{distro}";
+
+            
+        private static DirectoryInfo Extract(FileInfo input, DirectoryInfo output)
+        {
+            Console.WriteLine($"Extracting {input} to {output}");
+            output.Create();
+            using var inputStream = input.OpenRead();
+            using var reader = ReaderFactory.Open(inputStream);
+            reader.WriteAllToDirectory(output.FullName, new ExtractionOptions
+            {
+                ExtractFullPath = true,
+                Overwrite = true,
+
+                // throws NotImplementedException, need another way to preserve file permissions
+                // PreserveAttributes = true,
+            });
+
+            return new DirectoryInfo(Path.Join(output.FullName, "factorio"));
+        }
     }
 }
